init_code
