+++ config/csm/test_jmeter.yaml
@@ -34,12 +34,6 @@ test_44782:
   error_code: 4104
   message_id: "PermissionDenied"
   message_index: 2
-test_44794:
-  users_count : 5
-  buckets_count : 4
-  files_count : 65
-  rampup: 2
-  loop: 10
 test_44788:
   users_count : 1
   buckets_count : 1
